Fixes #33904 - Organizing fact parsers
Since parsers have own place in core, it's good to have them organized in better way. Current parsers are still in plugin-specific folders and it's not a good idea. So this PR introduces `FactParsers` namespace where parsers are much better organized. It also introduces `FactNames` namespace where is the same problem.
