+++ app/models/fact_names/puppet.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module FactNames
+  class Puppet < FactName
+    def origin
+      'Puppet'
+    end
+
+    def icon_path
+      "icons16x16/stub/goldenrod-p.png"
+    end
+  end
+end
