--- app/models/fact_names/puppet_fact_name.rb
@@ -1,9 +1,0 @@
-class PuppetFactName < FactName
-  def origin
-    'Puppet'
-  end
-
-  def icon_path
-    "icons16x16/stub/goldenrod-p.png"
-  end
-end
