+++ app/models/fact_names/salt.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
-module ForemanAnsible
-  # Define the class that fact names that come from Ansible should have
+module FactNames
+  # Define the class that fact names that come from Salt should have
   # It allows us to filter facts by origin, and also to display the origin
   # in the fact values table (/fact_values)
-  class FactName < ::FactName
+  class Salt < FactName
     def origin
-      'Ansible'
+      'Salt'
     end
 
     def icon_path
-      'Ansible.png'
+      'icons16x16/stub/black-s.png'
     end
   end
 end
