+++ config/initializers/plugin_parsers.rb
@@ -1,16 +1,16 @@
 Rails.application.config.to_prepare do
   # Puppet, the default parser
-  Foreman::Plugin.fact_parser_registry.register(:puppet, PuppetFactParser, true)
+  Foreman::Plugin.fact_parser_registry.register(:puppet, FactParsers::Puppet, true)
 
   # Ansible
-  Foreman::Plugin.fact_parser_registry.register(:ansible, AnsibleFactParser)
+  Foreman::Plugin.fact_parser_registry.register(:ansible, FactParsers::Ansible)
 
   # Katello
-  Foreman::Plugin.fact_parser_registry.register(Katello::RhsmFactName::FACT_TYPE, Katello::RhsmFactParser)
+  Foreman::Plugin.fact_parser_registry.register(FactNames::Rhsm::FACT_TYPE, FactParsers::Rhsm)
 
   # Chef
-  Foreman::Plugin.fact_parser_registry.register(:foreman_chef, ForemanChef::FactParser)
+  Foreman::Plugin.fact_parser_registry.register(:foreman_chef, FactParsers::Chef)
 
   # Salt
-  Foreman::Plugin.fact_parser_registry.register(:foreman_salt, ForemanSalt::FactParser)
+  Foreman::Plugin.fact_parser_registry.register(:foreman_salt, FactParsers::Salt)
 end
