+++ test/unit/foreman_ansible/fact_sparser_test.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'test_helper'
+
 module ForemanAnsible
   # Tests for checking if FactSparser can sparse a hash and unsparse it
   class FactSparserTest < ActiveSupport::TestCase
@@ -13,12 +15,12 @@ module ForemanAnsible
 
     test 'sparses simple hash' do
       assert_equal @sparsed_os_facts,
-        ForemanAnsible::FactSparser.sparse(@original_os_facts)
+        FactParsers::Utility::FactSparser.sparse(@original_os_facts)
     end
 
     test 'unsparse simple hash' do
       assert_equal @original_os_facts,
-        ForemanAnsible::FactSparser.unsparse(@sparsed_os_facts)
+        FactParsers::Utility::FactSparser.unsparse(@sparsed_os_facts)
     end
   end
 end
